requirements.txt
discord.py
yt_dlp
youtube-search-python
